Simplify and make speed improvements
Simplify gem so that we only have `#run` (or auto from command line) so
that we can undo some of the indirection in the hopes of speeding things
up.

One major change we make is to do the node matching in three separate
threads. See the changes in `match.rb`.