Add runner_spec
We add a runner_spec that though rough tests the core functionality of
the `#run` method. It checks that the runner correctly adds the `t`
methods in the `test.html.erb` file, and it checks that the correct
translations are added to the english locale. We also remove the
`ru.yml` file since translations for multiple languages was only part of
the interactive mode, which we removed.